This adds CMS support to the Ruby layer.
The CMS_ContentInfo object can be used.
Access to the CMS_SignedInfo structure exists, but it can not be created or freed,
as it is just a pointer into CMS_ContentInfo stack.
feat: added NOINTERN and NO_SIGNER_CERT_VERIFY flags
feat: look for CMS_sign, and set HAVE_CMS_SIGN if present, enabling CMS support